src/dlp-runner/requirements.txt
Flask==2.0.1
google-cloud-storage==1.41.1
google-cloud-dlp==3.2.2
gunicorn==20.1.0